Added schemas and all branches endpoints
